encapsulate components - Brief
