+++ src/components/Subset/Id.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Id = (props) => {
+    return (<>ID: <strong>{ props.children }</strong> </>);
+};
