+++ src/components/Subset/index.js
@@ -1,4 +1,6 @@
-export * from './SubsetPreview';
-export * from './SubsetBrief';
-export * from './Subsets';
-export * from './SubsetsBanner';
+export * from './Preview';
+export * from './Brief';
+export * from './BriefContextual';
+export * from './List';
+export * from './Banner';
+export * from './Id';
